test(NumIntegration): add missing error path test for ValueError on power of -1
This commit adds testing for the explicit ValueError exception raised when trying to integrate x^-1.
It explicitly validates the error message matching using pytest.raises.
Additionally, the tests for NumIntegration were extracted out of the monolithic test_Calculus.py into a dedicated test_NumIntegration.py module, removing duplicate tests and improving organization while maintaining 100% test coverage for the integration logic.
